requirements.txt
torch>=1.12.0
transformers>=4.15.0
datasets>=1.8.0
scikit-learn>=1.0.0
pyyaml>=5.4
tensorboard>=2.6.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.60.0
